2nd Commit
Started working on the entities system I'll be using. Decided that using a SimpleEntity Super class would be simplest to send to the client(s).

Figured out a system for converting and using 3d models.

As of right now entities are non functional. But hopefully in a week they will work perfectly.

Trying to move away from a SimplePhysicsInstance because sadly it contains too much data to send from server to client. to fix this I'm going to simplify so all moving objects in the games eye is considered an entity and treated as such.